+++ cassandra-test-image/src/main/docker/ecc-entrypoint.sh
@@ -86,4 +86,39 @@ EOF
 
 fi
 
+#
+# Setup nodetool SSL properties
+#
+  cat <<EOF > ~/.cassandra/nodetool-ssl.properties
+-Djavax.net.ssl.keyStore=/etc/certificates/.keystore
+-Djavax.net.ssl.keyStorePassword=ecctest
+-Djavax.net.ssl.trustStore=/etc/certificates/.truststore
+-Djavax.net.ssl.trustStorePassword=ecctest
+-Dcom.sun.management.jmxremote.ssl.need.client.auth=true
+EOF
+
+#
+# Setup nodetool status with SSL
+#
+  cat <<EOF > ~/.cassandra/nodetool-status-ssl.sh
+#!/bin/bash
+#
+# Copyright 2025 Telefonaktiebolaget LM Ericsson
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+nodetool --ssl -u cassandra -pw cassandra status
+EOF
+
+chmod 755 ~/.cassandra/nodetool-status-ssl.sh
+
 docker-entrypoint.sh