+++ standalone-integration/src/test/java/com/ericsson/bss/cassandra/ecchronos/standalone/ITOnDemandRepairJob.java
@@ -111,8 +111,7 @@ public class ITOnDemandRepairJob extends TestBase
     private Set<TableReference> myRepairs = new HashSet<>();
 
     @Parameterized.BeforeParam
-    public static void init(Boolean remoteRoutingOption) throws IOException
-    {
+    public static void init(Boolean remoteRoutingOption) throws IOException, InterruptedException {
         myRemoteRouting = remoteRoutingOption;
         initialize();